Settings: use .env location list for loading+
Load settings using a list of .env path locations so that it works
accross testing, containers and development. This is done since the path
to .env changes between these environments. The path to .env is loaded
by priority in the same order that they are in the list, the first path
found is loaded.